refactor Navbar
