+++ campus-connect-app/app/components/Navbar.tsx
@@ -1,0 +1,209 @@
+import {
+  Disclosure,
+  DisclosureButton,
+  DisclosurePanel,
+  Menu,
+  MenuButton,
+  MenuItem,
+  MenuItems,
+} from "@headlessui/react";
+import { MagnifyingGlassIcon } from "@heroicons/react/20/solid";
+import { Bars3Icon, BellIcon, XMarkIcon } from "@heroicons/react/24/outline";
+import Link from "next/link";
+
+const user = {
+  name: "Tom Cook",
+  email: "tom@example.com",
+  imageUrl:
+    "https://images.unsplash.com/photo-1472099645785-5658abf4ff4e?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=facearea&facepad=2&w=256&h=256&q=80",
+};
+const navigation = [
+  { name: "Dashboard", href: "/dashboard", current: true },
+  { name: "Clubs", href: "/clubs", current: false },
+  { name: "Events", href: "/events", current: false },
+];
+const userNavigation = [
+  { name: "Your Profile", href: "#" },
+  { name: "Settings", href: "#" },
+  { name: "Sign out", href: "#" },
+];
+
+function classNames(...classes) {
+  return classes.filter(Boolean).join(" ");
+}
+
+export default function Navbar() {
+  return (
+    <>
+      <Disclosure
+        as="nav"
+        className="border-b border-indigo-300/25 bg-indigo-600 lg:border-none"
+      >
+        <div className="mx-auto max-w-7xl px-2 sm:px-4 lg:px-8">
+          <div className="relative flex h-16 items-center justify-between lg:border-b lg:border-indigo-400/25">
+            <div className="flex items-center px-2 lg:px-0">
+              <div className="shrink-0">
+                <img
+                  alt="Your Company"
+                  src="https://tailwindcss.com/plus-assets/img/logos/mark.svg?color=indigo&shade=300"
+                  className="block size-8"
+                />
+              </div>
+              <div className="hidden lg:ml-10 lg:block">
+                <div className="flex space-x-4">
+                  {navigation.map((item) => (
+                    <Link
+                      key={item.name}
+                      href={item.href}
+                      aria-current={item.current ? "page" : undefined}
+                      className={classNames(
+                        item.current
+                          ? "bg-indigo-700 text-white"
+                          : "text-white hover:bg-indigo-500/75",
+                        "rounded-md px-3 py-2 text-sm font-medium"
+                      )}
+                    >
+                      {item.name}
+                    </Link>
+                  ))}
+                </div>
+              </div>
+            </div>
+            <div className="flex flex-1 justify-center px-2 lg:ml-6 lg:justify-end">
+              <div className="grid w-full max-w-lg grid-cols-1 lg:max-w-xs">
+                <input
+                  name="search"
+                  type="search"
+                  placeholder="Search"
+                  aria-label="Search"
+                  className="col-start-1 row-start-1 block w-full rounded-md bg-white py-1.5 pr-3 pl-10 text-base text-gray-900 outline-hidden placeholder:text-gray-400 focus:outline-2 focus:outline-offset-2 focus:outline-white/40 sm:text-sm/6"
+                />
+                <MagnifyingGlassIcon
+                  aria-hidden="true"
+                  className="pointer-events-none col-start-1 row-start-1 ml-3 size-5 self-center text-gray-400"
+                />
+              </div>
+            </div>
+            <div className="flex lg:hidden">
+              {/* Mobile menu button */}
+              <DisclosureButton className="group relative inline-flex items-center justify-center rounded-md bg-indigo-600 p-2 text-indigo-200 hover:bg-indigo-500/75 hover:text-white focus:ring-2 focus:ring-white focus:ring-offset-2 focus:ring-offset-indigo-600 focus:outline-hidden">
+                <span className="absolute -inset-0.5" />
+                <span className="sr-only">Open main menu</span>
+                <Bars3Icon
+                  aria-hidden="true"
+                  className="block size-6 group-data-open:hidden"
+                />
+                <XMarkIcon
+                  aria-hidden="true"
+                  className="hidden size-6 group-data-open:block"
+                />
+              </DisclosureButton>
+            </div>
+            <div className="hidden lg:ml-4 lg:block">
+              <div className="flex items-center">
+                <button
+                  type="button"
+                  className="relative shrink-0 rounded-full bg-indigo-600 p-1 text-indigo-200 hover:text-white focus:ring-2 focus:ring-white focus:ring-offset-2 focus:ring-offset-indigo-600 focus:outline-hidden"
+                >
+                  <span className="absolute -inset-1.5" />
+                  <span className="sr-only">View notifications</span>
+                  <BellIcon aria-hidden="true" className="size-6" />
+                </button>
+
+                {/* Profile dropdown */}
+                <Menu as="div" className="relative ml-3 shrink-0">
+                  <div>
+                    <MenuButton className="relative flex rounded-full bg-indigo-600 text-sm text-white focus:ring-2 focus:ring-white focus:ring-offset-2 focus:ring-offset-indigo-600 focus:outline-hidden">
+                      <span className="absolute -inset-1.5" />
+                      <span className="sr-only">Open user menu</span>
+                      <img
+                        alt=""
+                        src={user.imageUrl}
+                        className="size-8 rounded-full"
+                      />
+                    </MenuButton>
+                  </div>
+                  <MenuItems
+                    transition
+                    className="absolute right-0 z-10 mt-2 w-48 origin-top-right rounded-md bg-white py-1 ring-1 shadow-lg ring-black/5 transition focus:outline-hidden data-closed:scale-95 data-closed:transform data-closed:opacity-0 data-enter:duration-100 data-enter:ease-out data-leave:duration-75 data-leave:ease-in"
+                  >
+                    {userNavigation.map((item) => (
+                      <MenuItem key={item.name}>
+                        <Link
+                          href={item.href}
+                          className="block px-4 py-2 text-sm text-gray-700 data-focus:bg-gray-100 data-focus:outline-hidden"
+                        >
+                          {item.name}
+                        </Link>
+                      </MenuItem>
+                    ))}
+                  </MenuItems>
+                </Menu>
+              </div>
+            </div>
+          </div>
+        </div>
+
+        <DisclosurePanel className="lg:hidden">
+          <div className="space-y-1 px-2 pt-2 pb-3">
+            {navigation.map((item) => (
+              <DisclosureButton
+                key={item.name}
+                as="a"
+                href={item.href}
+                aria-current={item.current ? "page" : undefined}
+                className={classNames(
+                  item.current
+                    ? "bg-indigo-700 text-white"
+                    : "text-white hover:bg-indigo-500/75",
+                  "block rounded-md px-3 py-2 text-base font-medium"
+                )}
+              >
+                {item.name}
+              </DisclosureButton>
+            ))}
+          </div>
+          <div className="border-t border-indigo-700 pt-4 pb-3">
+            <div className="flex items-center px-5">
+              <div className="shrink-0">
+                <img
+                  alt=""
+                  src={user.imageUrl}
+                  className="size-10 rounded-full"
+                />
+              </div>
+              <div className="ml-3">
+                <div className="text-base font-medium text-white">
+                  {user.name}
+                </div>
+                <div className="text-sm font-medium text-indigo-300">
+                  {user.email}
+                </div>
+              </div>
+              <button
+                type="button"
+                className="relative ml-auto shrink-0 rounded-full bg-indigo-600 p-1 text-indigo-200 hover:text-white focus:ring-2 focus:ring-white focus:ring-offset-2 focus:ring-offset-indigo-600 focus:outline-hidden"
+              >
+                <span className="absolute -inset-1.5" />
+                <span className="sr-only">View notifications</span>
+                <BellIcon aria-hidden="true" className="size-6" />
+              </button>
+            </div>
+            <div className="mt-3 space-y-1 px-2">
+              {userNavigation.map((item) => (
+                <DisclosureButton
+                  key={item.name}
+                  as="a"
+                  href={item.href}
+                  className="block rounded-md px-3 py-2 text-base font-medium text-white hover:bg-indigo-500/75"
+                >
+                  {item.name}
+                </DisclosureButton>
+              ))}
+            </div>
+          </div>
+        </DisclosurePanel>
+      </Disclosure>
+    </>
+  );
+}
